Add generator entrypoint in CLI for aiida workflow generation
This PR uses the work in MR
[!300](https://earth.bsc.es/gitlab/es/autosubmit/-/merge_requests/300)
as base to add an entry point to the CLI to generate from autosubmit
workflows an AiiDA version. I take a bit different approach than MR
[!300](https://earth.bsc.es/gitlab/es/autosubmit/-/merge_requests/300)
in the sense that no AiiDA object is generated in autosubmit but a
completely self-contained folder with all scripts and files together
with a script that runs the AiiDA workflow. This has the advantage that
the aiida workflow can be generated without having aiida installed.